cleanup and reorganise
